changed architecture project
web main.go now lies in src/
