--- cmd/lab1/main.go
@@ -1,14 +1,0 @@
-package main
-
-import (
-	"log"
-
-	"test/internal/api"
-)
-
-
-func main() {
-	log.Println("App start")
-	api.StartServer()
-	log.Println("App down")
-}
