+++ src/backend/db.go
@@ -1,4 +1,4 @@
-package api
+package backend
 
 import (
 	"strings"
@@ -327,25 +327,6 @@ func FilterLangsByQuery(langs []Lang, query string) []Lang {
 		}
 	}
 	return filteredLangs
-}
-
-// Парсинг списка в карту
-func ParseList(listStr string) map[string]string {
-	result := make(map[string]string)
-	lines := strings.Split(listStr, "\n")
-
-	for _, line := range lines {
-		line = strings.TrimSpace(line)
-		if line == "" || !strings.Contains(line, ":") {
-			continue
-		}
-
-		parts := strings.SplitN(line, ":", 2)
-		key := strings.TrimSpace(parts[0]) + ":"
-		value := strings.TrimSpace(parts[1])
-		result[key] = value
-	}
-	return result
 }
 
 // func FindMaxProjectID() int {
