+++ src/backend/utils.go
@@ -1,0 +1,24 @@
+package backend
+
+import (
+	"strings"
+)
+
+// Парсинг списка в карту
+func ParseList(listStr string) map[string]string {
+	result := make(map[string]string)
+	lines := strings.Split(listStr, "\n")
+
+	for _, line := range lines {
+		line = strings.TrimSpace(line)
+		if line == "" || !strings.Contains(line, ":") {
+			continue
+		}
+
+		parts := strings.SplitN(line, ":", 2)
+		key := strings.TrimSpace(parts[0]) + ":"
+		value := strings.TrimSpace(parts[1])
+		result[key] = value
+	}
+	return result
+}
