+++ static/css/information.css
@@ -43,6 +43,17 @@ body {
   border-radius: 5px;
 }
 
+.info {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  max-width: 1940px;
+  margin: 0 auto;
+  position: relative;
+  padding: 0;
+  box-sizing: border-box;
+}
+
 .icon-btn-back {
   width: 50px;
   height: 50px;
