+++ templates/applications.tmpl
@@ -12,7 +12,7 @@
     <div class="writing-code">
       <div class="header">
         <a href="/home" class="background-btn-back">
-            <img class="icon-btn-back" src="/image/icon-btn-back.png" />
+            <img class="icon-btn-back" src="/static/img/icon-btn-back.png" />
         </a>
       </div>
       <div class="tree">
