+++ templates/services.tmpl
@@ -18,7 +18,7 @@
   <body>
     <div class="services-container">
       <a href="/home" class="background-btn-back">
-        <img class="icon-btn-back" src="/image/icon-btn-back.png" />
+        <img class="icon-btn-back" src="/static/img/icon-btn-back.png" />
       </a>
       <div class="header">
         <div class="search-section">
@@ -31,16 +31,16 @@
               placeholder="Поиск..."
             />
             <button type="submit" style="display: none;">Поиск</button>
-            <img class="search-icon" src="image/icon-find.png" />
+            <img class="search-icon" src="/static/img/icon-find.png" />
           </form>
         </div>
         <div class="file-count-section">
           <a href="/app/1" class="file-count">
-            <img class="file-count-icon" src="image/icon-count-files.png" />
+            <img class="file-count-icon" src="/static/img/icon-count-files.png" />
             <div class="file-count-text">3</div>
           </a>
         </div>
-        <img class="separator-line" src="image/line.png" />
+        <img class="separator-line" src="/static/img/line.png" />
       </div>
       <ul class="service-list two-columns">
         {{ range .Langs }}
